# valo-mapper-api/docs/swagger.yaml
basePath: /
definitions:
  handlers.BillingPlanPriceResponse:
    properties:
      currency:
        type: string
      interval:
        type: string
      intervalCount:
        type: integer
      plan:
        type: string
      priceId:
        type: string
      unitAmount:
        type: integer
      unitAmountDecimal:
        type: string
    type: object
  handlers.BillingPlansResponse:
    properties:
      monthly:
        $ref: '#/definitions/handlers.BillingPlanPriceResponse'
      yearly:
        $ref: '#/definitions/handlers.BillingPlanPriceResponse'
    type: object
...
  handlers.CreateCheckoutSessionRequest:
    properties:
      plan:
        type: string
      returnTo:
        type: string
    type: object
  handlers.CreateCheckoutSessionResponse:
    properties:
      sessionId:
        type: string
      url:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
      requestId:
        type: string
    type: object
...
info:
  contact: {}
  description: API for ValoMapper backend services.
  title: ValoMapper API
  version: "1.0"
paths:
...
  /api/billing/checkout-session:
    post:
      consumes:
      - application/json
      description: Creates a Stripe checkout session for a monthly or yearly plan.
      parameters:
      - description: Checkout request
        in: body
        name: request
        schema:
          $ref: '#/definitions/handlers.CreateCheckoutSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CreateCheckoutSessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Stripe checkout session
      tags:
      - billing
  /api/billing/plans:
    get:
      description: Returns Stripe-backed monthly and yearly pricing details.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BillingPlansResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get billing plans
      tags:
      - billing
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Firebase bearer token in the form "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
  InternalAPIKey:
    description: Internal service API key.
    in: header
    name: X-Internal-API-Key
    type: apiKey
swagger: "2.0"